separation of schema and data validations
