+++ app/api/entities/specs/entities.spec.js
@@ -665,6 +665,7 @@ describe('entities', () => {
         })
         .catch(catchErrors(done));
     });
+
     it('should sanitize daterange, removing non valid dates', done => {
       const doc1 = {
         _id: batmanFinishesId,
